Create Rust listener example
To return a boolean value from a Rust dApp to a Solidity smart contract, you will need to use the web3 Rust library to connect to an Ethereum node and send a transaction to the smart contract with the desired boolean value as the input.

In this example, we first define the ABI for the smart contract, which specifies the functions and their inputs and outputs. We then connect to an Ethereum node using the web3 library and use the ABI to create a Contract instance.

We then use the setBoolValue function of the contract to set the boolean value to true, and send the transaction to the Ethereum network. Once the transaction is mined, we use the boolValue function to retrieve the current boolean value and check that it is equal to true.